# starter-frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Starter API
  version: 1.0.0
paths:
...
  /api/user/me/:
    get:
      operationId: user_me_retrieve
      summary: Récupérer et mettre à jour mon profil
      tags:
      - User
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomUserRead'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Validation error
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Forbidden
    patch:
      operationId: user_me_partial_update
      summary: Récupérer et mettre à jour mon profil
      tags:
      - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCustomUserProfileUpdateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCustomUserProfileUpdateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCustomUserProfileUpdateRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomUserRead'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Validation error
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Forbidden
...
components:
  schemas:
...
    CustomUserRead:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          readOnly: true
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
        email:
          type: string
          format: email
          nullable: true
          title: Email address
          maxLength: 254
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        language:
          $ref: '#/components/schemas/LanguageEnum'
        email_confirmed:
          type: boolean
          readOnly: true
        password_change_required:
          type: boolean
          readOnly: true
        is_superuser:
          type: boolean
          readOnly: true
          title: Superuser status
          description: Designates that this user has all permissions without explicitly
            assigning them.
        is_staff:
          type: boolean
          readOnly: true
          title: Staff status
          description: Designates whether the user can log into this admin site.
        is_active:
          type: boolean
          readOnly: true
          title: Active
          description: Designates whether this user should be treated as active. Unselect
            this instead of deleting accounts.
      required:
      - email_confirmed
      - id
      - is_active
      - is_staff
      - is_superuser
      - password_change_required
      - username
...
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
...
    LanguageEnum:
      enum:
      - en
      - fr
      - nl
      - it
      - es
      type: string
      description: |-
        * `en` - English
        * `fr` - French
        * `nl` - Dutch
        * `it` - Italy
        * `es` - Spain
...
    PatchedCustomUserProfileUpdateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          nullable: true
          title: Email address
          maxLength: 254
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        language:
          $ref: '#/components/schemas/LanguageEnum'
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT